+++ jsdp/src/main/java/jsdp/app/routing/location/BRL_ForwardRecursion.java
@@ -53,7 +53,7 @@ public class BRL_ForwardRecursion extends ForwardRecursion {
                                float loadFactor,
                                SamplingScheme samplingScheme,
                                int sampleSize,
-                               int reductionFactorPerStage){
+                               double reductionFactorPerStage){
       super(OptimisationDirection.MIN);
       this.horizonLength = horizonLength;
       this.machineLocation = machineLocation;
